requirements.txt
Django==4.2.29
djangorestframework==3.16.1
django-autoslug==1.9.9
django-filter==25.1
psycopg2-binary==2.9.11
celery==5.6.2
redis==6.4.0
gunicorn==23.0.0
asgiref==3.11.1